# aggregate: directly hash the list values