--- a/query/physicalplan/aggregate.go
+++ b/query/physicalplan/aggregate.go
@@ -12,6 +12,7 @@ import (
 	"github.com/apache/arrow/go/v13/arrow/math"
 	"github.com/apache/arrow/go/v13/arrow/memory"
 	"github.com/apache/arrow/go/v13/arrow/scalar"
+	"github.com/cespare/xxhash/v2"
 	"github.com/dgryski/go-metro"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
@@ -281,18 +282,20 @@ func hashArray(arr arrow.Array) []uint64 {
 
 func hashListArray(arr *array.List) []uint64 {
 	res := make([]uint64, arr.Len())
+	digest := xxhash.New()
 	for i := 0; i < arr.Len(); i++ {
-		list := []byte{}
 		if err := arrowutils.ForEachValueInList(i, arr, func(_ int, v any) {
 			switch val := v.(type) {
 			case []byte:
-				list = append(list, val...)
+				// NOTE: We can ignore the return here because from the xxhash Write function is says: It always returns len(b), nil.
+				_, _ = digest.Write(val)
 			}
 		}); err != nil {
 			panic(err)
 		}
 
-		res[i] = metro.Hash64(list, 0)
+		res[i] = digest.Sum64()
+		digest.Reset()
 	}
 	return res
 }