Previously we were appending all the list values and then hashing them. This becomes extremely inefficient when using really large lists. Instead we just hash the list values directly.

Below is a benchmark for a normal aggregate query against 12h of real profiling data.

```bash
name              old time/op    new time/op    delta
_PreAggregate-10     6.56s ± 2%     4.81s ± 3%  -26.76%  (p=0.000 n=10+10)

name              old alloc/op   new alloc/op   delta
_PreAggregate-10    33.1GB ± 0%    14.6GB ± 0%  -55.94%  (p=0.000 n=8+10)

name              old allocs/op  new allocs/op  delta
_PreAggregate-10      174M ± 0%      162M ± 0%   -6.99%  (p=0.000 n=10+9)
```